agents/viz-agent/pyproject.toml
[project]
name = "ai-cfo-viz-agent"
version = "0.1.0"
description = "Visualization Agent for AI CFO BI Agent"
authors = [
    {name = "AI CFO Team", email = "team@ai-cfo.com"}
]
dependencies = [
    "pydantic==2.5.0",
    "redis==5.0.1",
    "pika==1.3.2",
    "plotly==5.17.0",
    "matplotlib==3.8.2",
    "pandas==2.1.4",
    "numpy==1.26.2",
    "python-dotenv==1.0.0",
    "orjson==3.9.10",
    "structlog==23.2.0",
    "reportlab==4.0.7",
    "openpyxl==3.1.2",
    "psutil==5.9.6",
    "kaleido==0.2.1",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88